requirements.txt
fastapi
uvicorn
jinja2
openpyxl
python-multipart
playwright
bs4
python-crontab
httpx
beautifulsoup4
pydantic
starlette
SQLAlchemy
greenlet
pillow
pytesseract
PyPDF2<3.0
pdf2image
poppler-utils
passlib[bcrypt]
python-jose[cryptography]
